New configuration parameter
Let the user decide whether she or he wants the algorithms for detecting
relevant code changes to consider all code blocks or only those that
reference a configuration option of the variability model.
Hence, we introduce the new and optional property
"analysis.dead_code_change_analyzer.consider_all_blocks", which takes
"true" or "false" as values. The default value, e.g., if not set, is
"true".