Revise CodeFileDiff completely
Due to the fact that we have to develop something like a change-aware
code block parser to detect all changes and, in particular, their impact
correctly, we decided to revise the code file diff class. It now only
detects changes to block informations without further consideration of
whether there are references to configuration options. This still leads
to the same results as it detects all block changes as being relevant.
However, a detailed analysis of the impact is not performed anymore.
This requires checks based on the presence condition, which are much
easier to achieve by using a code block extractor.